requirements.txt
audioread
beautifulsoup4
bs4
fastapi
fastembed
FlashRank
google-auth
google-genai
hf-xet
html2text
huggingface_hub
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
librosa
numpy
pillow
pycparser
pydantic
pydantic-settings
pydantic_core
python-dotenv
python-multipart
PyYAML
qdrant-client
regex
requests
requests-toolbelt
rsa
safetensors
sarvamai
scikit-learn
scipy
sentence-transformers
setuptools
soundfile
soupsieve
tenacity
threadpoolctl
tokenizers
torch
torchaudio
tqdm
transformers
urllib3
uuid_utils
uvicorn
websockets
wikipedia
